Add neovim_nightly image
Add image that contains the latest nightly build of neovim with my
configuration and plugins.

The neovim_nightly.def is the main file and the one that should be used
(using Alpine for a smaller image).  The other is based on Ubuntu
(easier to set up but larger) and is only kept for reference in case I
run into unsolvable issues with Alpine at some point.